requirements.txt
numpy==1.19.5
pygame==2.0.0
pyserial==3.4
networkx==2.5